tests: connectivity: proxy: Add sleep before checking logs
Lately we have seen this test fail because at the second http-connect
test we would still fetch the old socks5 logs from the glider container.
So let's add a 3 seconds sleep in between the api ping call and the log
check. If even after this 3 seconds sleep we get the old logs then it's
an indication of a problem with either the container being slow to run
or a problem with slow logs output from the glider container.

Change-type: patch
Signed-off-by: Florin Sarbu <florin@balena.io>